+++ scripts/scripts.js
@@ -13,7 +13,7 @@ import {
   loadCSS,
 } from './lib-franklin.js';
 
-import { decorateAsFluidwebPage, decorateMedia } from './fluidweb.js';
+import { decorateAsFluidwebPage, decorateMedia } from './lib-fluidweb.js';
 
 const LCP_BLOCKS = []; // add your LCP blocks to the list
 window.hlx.RUM_GENERATION = 'project-1'; // add your RUM generation information here
